Extract pipes validations
